% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biADMM.R
\name{biADMM}
\alias{biADMM}
\title{Algorithms for Convex Biclustering}
\usage{
biADMM(
  X,
  nu1,
  nu2,
  gamma_1,
  gamma_2,
  kk = 5,
  phi = 0.5,
  niter = 1000,
  tol = 0.1,
  output = 1
)
}
\arguments{
\item{output}{}
}
\value{

}
\description{
Algorithms for Convex Biclustering
}
\examples{
# generate dataset
set.seed(123)
X = data_gen(n = 100, p = 80)
# set parameters
nu1 = nu2 = gamma_1 = gamma_2 = 0.1
kk = 5
phi = 0.5
# biADMM algorithm
res1 = biADMM(X, nu1, nu2, gamma_1, gamma_2,
 kk, phi, niter = 1000, tol = 0.0001, output = 0)
}
