% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biADMM.speed.R
\name{biADMM.speed}
\alias{biADMM.speed}
\title{Algorithms for Convex Biclustering}
\usage{
biADMM.speed(
  X,
  nu1,
  nu2,
  gamma_1,
  gamma_2,
  kk,
  phi,
  niters = 10,
  tol = 0.1,
  output = 1
)
}
\arguments{
\item{output}{}
}
\value{

}
\description{
Algorithms for Convex Biclustering
}
\examples{



}
