% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{data_gen}
\alias{data_gen}
\title{Simulated data set generation}
\usage{
data_gen(
  seed.cluster = 123,
  seed.data = 654,
  n,
  p,
  theta = 2,
  row_group = 4,
  col_group = 4
)
}
\arguments{
\item{seed.cluster}{seed.cluster is used to control the cluster assignment}

\item{seed.data}{seed.data is used to control the data generation given clustering structure}

\item{n}{number of subjects}

\item{p}{number of features}

\item{theta}{the variance of the x, which is sampled from a normal distribution}

\item{row_group}{number of row clusters}

\item{col_group}{number of column clusters}
}
\value{
Output is the simulated dataset
}
\description{
\code{data_gen} generates datasets in the simulation study.
}
\examples{
set.seed(123)
data = data_gen(n = 100, p = 80)
}
