% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{elk}
\alias{elk}
\title{Calculation of e matrices}
\usage{
elk(n, p)
}
\arguments{
\item{n}{the number of rows}

\item{p}{the numbwr of columns}
}
\value{
A list of e matrices.
\cite{el1} is the matrix of el1 in the paper, which is constructed by column matrix ei. ei is an n-dimensional vector with each component being 0 but its i-th component being 1.

\cite{el2} is the matrix of el2 in the paper, which is constructed by column matrix ei. ei is an n-dimensional vector with each component being 0 but its i-th component being 1.

\cite{ek1} is the matrix of ek1 in the paper, which is constructed by column matrix ei. ei is an p-dimensional vector with each component being 0 but its i-th component being 1.

\cite{ek2} is the matrix of ek2 in the paper, which is constructed by column matrix ei. ei is an p-dimensional vector with each component being 0 but its i-th component being 1.
}
\description{
\code{elk} generates the e matrices.
}
\examples{
elk(4,3)
}
