% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sylvester.R
\name{sylvester}
\alias{sylvester}
\title{Solve Sylvester equation AX + XB = C.}
\usage{
sylvester(A, B, C, tol = 1e-04)
}
\arguments{
\item{A}{Input matrix}

\item{B}{Input matrix}

\item{C}{Input matrix}

\item{tol}{The tolerance of the algorithm}
}
\value{
X Output matrix
}
\description{
\code{sylvester} finds the matrix X that obey this equation, given matrices A, B, and C.
}
\examples{
set.seed(123)
A <- matrix(rnorm(9),3,3)
B <- matrix(rnorm(9),3,3)
Xtrue <- matrix(rnorm(9),3,3)
C <- A \%*\% Xtrue + Xtrue \%*\% B
X <- sylvester(A,B,C)
X - Xtrue
}
