% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biADMM.speed.R
\name{biADMM.speed}
\alias{biADMM.speed}
\title{bi-ADMM: a Biclustering Algorithm for the General Model (faster version)}
\usage{
biADMM.speed(
  X,
  nu1,
  nu2,
  gamma_1,
  gamma_2,
  m,
  phi,
  prox = "l2",
  niters = 10,
  tol = 0.1,
  output = 1
)
}
\arguments{
\item{X}{The data matrix to be clustered. The rows are the samples, and the columns are the features.}

\item{nu1}{A regularization parameter for row shrinkage}

\item{nu2}{A regularization parameter for column shrinkage}

\item{gamma_1}{A regularization parameter for row shrinkage}

\item{gamma_2}{A regularization parameter for column shrinkage}

\item{m}{m-nearest-neighbors in the weight function}

\item{phi}{The parameter phi in the weight function}

\item{prox}{The proximal maps. Could calculate L1 norm, L2 norm, or L-infinity, use "l1", "l2", or "l-inf", respectively.}

\item{niters}{Iteraion times}

\item{tol}{Stopping criterion}

\item{output}{When output = 1, print the results at each iteration. No print when output equals other value.}
}
\value{
A list of results, containing matrix of A, v, z, lambda1, and lambda2
}
\description{
Same algorithm as \code{biADMM}. Call python code to speed up the running time.
}
\examples{
# generate dataset
set.seed(123)
X = data_gen(n = 100, p = 80)
# set parameters
nu1 = nu2 = gamma_1 = gamma_2 = 0.1
m = 5
phi = 0.5
# biADMM algorithm
res2 = biADMM.speed(X, nu1, nu2, gamma_1, gamma_2,
 m, phi, niter = 10, tol = 0.0001, output = 0)
dim(res2$A)
}
